requirements.txt
asgiref==3.8.1
Brotli==1.1.0
Django==5.2
gunicorn==23.0.0
packaging==25.0
sqlparse==0.5.3
whitenoise==6.9.0

# AI puzzle generation
anthropic>=0.39.0
openai>=1.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
httpx>=0.25.0